requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
lxml==6.0.2
PyMuPDF==1.26.5
psycopg[binary]
psycopg2==2.9.12
python-docx==1.2.0
python-dotenv
pywin32==311; platform_system == "Windows"
pyhwp==0.1b15
requests==2.33.0
schedule==1.2.2
sentence-transformers
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
